Create Project Euler 124.lisp
I used a modified sieve of Eratosthenes. I started by filling out an array of 1's, traversing it, and, instead of crossing out the multiples of each prime, I multiply the numbers in the positions corresponding to multiples of the prime by the prime. E.g.

1 1 1 1 1 1 1 1 ...
1 2 1 2 1 2 1 2 ...
1 2 3 2 1 6 1 2 ...
1 2 3 2 5 6 1 2 ...

With the list of radicals, sorting and finding the desired number is trivial.